requirements.txt
asgiref==3.8.1
channels==4.1.0
cloudinary==1.39.1
cryptography==41.0.0
daphne==4.1.2
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
django-rest-passwordreset==1.3.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
